# backend/pkg/server/docs/swagger.yaml
basePath: /api/v1
definitions:
  ErrorResponse:
    properties:
      code:
        example: Internal
        type: string
      error:
        example: original server error message
        type: string
      msg:
        example: internal server error
        type: string
      status:
        example: error
        type: string
    type: object
  SuccessResponse:
    properties:
      data:
        type: object
      status:
        example: success
        type: string
    type: object
...
  models.APIToken:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        minimum: 0
        type: integer
      name:
        maxLength: 100
        type: string
      role_id:
        minimum: 0
        type: integer
      status:
        type: string
      token_id:
        type: string
      ttl:
        maximum: 94608000
        minimum: 60
        type: integer
      updated_at:
        type: string
      user_id:
        minimum: 0
        type: integer
    required:
    - created_at
    - status
    - token_id
    - ttl
    - updated_at
    type: object
  models.APITokenWithSecret:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        minimum: 0
        type: integer
      name:
        maxLength: 100
        type: string
      role_id:
        minimum: 0
        type: integer
      status:
        type: string
      token:
        type: string
      token_id:
        type: string
      ttl:
        maximum: 94608000
        minimum: 60
        type: integer
      updated_at:
        type: string
      user_id:
        minimum: 0
        type: integer
    required:
    - created_at
    - status
    - token
    - token_id
    - ttl
    - updated_at
    type: object
...
  models.CreateAPITokenRequest:
    properties:
      name:
        maxLength: 100
        type: string
      ttl:
        description: from 1 minute to 3 years
        maximum: 94608000
        minimum: 60
        type: integer
    required:
    - ttl
    type: object
...
  models.UpdateAPITokenRequest:
    properties:
      name:
        maxLength: 100
        type: string
      status:
        type: string
    type: object
...
  services.tokens:
    properties:
      tokens:
        items:
          $ref: '#/definitions/models.APIToken'
        type: array
      total:
        type: integer
    type: object
...
info:
  contact:
    email: team@pennull.com
    name: penNULL Development Team
    url: https://pennull.com
  description: Swagger API for Penetration Testing Advanced General Intelligence penNULL.
  license:
    name: MIT
    url: https://opensource.org/license/mit
  termsOfService: http://swagger.io/terms/
  title: penNULL Swagger API
  version: "1.0"
paths:
...
  /tokens:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: tokens retrieved successful
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/services.tokens'
              type: object
        "403":
          description: listing tokens not permitted
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: internal error on listing tokens
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: List API tokens
      tags:
      - Tokens
    post:
      consumes:
      - application/json
      parameters:
      - description: Token creation request
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/models.CreateAPITokenRequest'
      produces:
      - application/json
      responses:
        "201":
          description: token created successful
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.APITokenWithSecret'
              type: object
        "400":
          description: invalid token request or default salt
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: creating token not permitted
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: internal error on creating token
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Create new API token for automation
      tags:
      - Tokens
  /tokens/{tokenID}:
    delete:
      parameters:
      - description: Token ID
        in: path
        name: tokenID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: token deleted successful
          schema:
            $ref: '#/definitions/SuccessResponse'
        "403":
          description: deleting token not permitted
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: token not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: internal error on deleting token
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Delete API token
      tags:
      - Tokens
    get:
      parameters:
      - description: Token ID
        in: path
        name: tokenID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: token retrieved successful
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.APIToken'
              type: object
        "403":
          description: accessing token not permitted
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: token not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: internal error on getting token
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get API token details
      tags:
      - Tokens
    put:
      consumes:
      - application/json
      parameters:
      - description: Token ID
        in: path
        name: tokenID
        required: true
        type: string
      - description: Token update request
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/models.UpdateAPITokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: token updated successful
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.APIToken'
              type: object
        "400":
          description: invalid update request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: updating token not permitted
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: token not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: internal error on updating token
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Update API token
      tags:
      - Tokens
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"